move logic to a run.py file
